add/update doc-stubs
