+++ stubs/micropython-v1_17-docstubs/array.pyi
@@ -1,6 +1,1 @@
-from typing import Any, Optional
-
-class array:
-    def __init__(self, typecode, iterable: Optional[Any]) -> None: ...
-    def append(self, val) -> Any: ...
-    def extend(self, iterable) -> Any: ...
+from uarray import *
