+++ stubs/micropython-v1_17-docstubs/binascii.pyi
@@ -1,6 +1,1 @@
-from typing import Any, Optional
-
-def hexlify(data, sep: Optional[Any]) -> bytes: ...
-def unhexlify(data) -> bytes: ...
-def a2b_base64(data) -> bytes: ...
-def b2a_base64(data) -> bytes: ...
+from ubinascii import *
