+++ stubs/micropython-v1_17-docstubs/json.pyi
@@ -1,6 +1,1 @@
-from typing import Any
-
-def dump(obj, stream, separators: Any | None = ...) -> Any: ...
-def dumps(obj, separators: Any | None = ...) -> str: ...
-def load(stream) -> Any: ...
-def loads(str) -> Any: ...
+from ujson import *
