+++ stubs/micropython-v1_17-docstubs/modules.json
@@ -153,12 +153,56 @@
             "module": "time"
         },
         {
+            "file": "uarray.py",
+            "module": "uarray"
+        },
+        {
             "file": "uasyncio.py",
             "module": "uasyncio"
         },
         {
+            "file": "ubinascii.py",
+            "module": "ubinascii"
+        },
+        {
             "file": "uctypes.py",
             "module": "uctypes"
+        },
+        {
+            "file": "uio.py",
+            "module": "uio"
+        },
+        {
+            "file": "ujson.py",
+            "module": "ujson"
+        },
+        {
+            "file": "uos.py",
+            "module": "uos"
+        },
+        {
+            "file": "uselect.py",
+            "module": "uselect"
+        },
+        {
+            "file": "usocket.py",
+            "module": "usocket"
+        },
+        {
+            "file": "ussl.py",
+            "module": "ussl"
+        },
+        {
+            "file": "ustruct.py",
+            "module": "ustruct"
+        },
+        {
+            "file": "utime.py",
+            "module": "utime"
+        },
+        {
+            "file": "uzlib.py",
+            "module": "uzlib"
         },
         {
             "file": "wipy.py",
