+++ stubs/micropython-v1_17-docstubs/network.py
@@ -297,16 +297,16 @@ class WLANWiPy(AbstractNIC):
     Create a WLAN object, and optionally configure it. See `init()` for params of configuration.
     """
 
-    STA: Any
+    STA: Any = ...
     #    selects the WLAN mode
-    AP: Any
-    WEP: Any
-    WPA: Any
+    AP: Any = ...
+    WEP: Any = ...
+    WPA: Any = ...
     #    selects the network security
-    WPA2: Any
-    INT_ANT: Any
+    WPA2: Any = ...
+    INT_ANT: Any = ...
     #    selects the antenna type
-    EXT_ANT: Any
+    EXT_ANT: Any = ...
 
     def __init__(self, id=0, *args) -> None:
         ...
@@ -455,10 +455,10 @@ class CC3K:
       nic = network.CC3K(pyb.SPI(2), pyb.Pin.board.Y5, pyb.Pin.board.Y4, pyb.Pin.board.Y3)
     """
 
-    WEP: Any
-    WPA: Any
+    WEP: Any = ...
+    WPA: Any = ...
     #    security type to use
-    WPA2: Any
+    WPA2: Any = ...
 
     def __init__(self, spi, pin_cs, pin_en, pin_irq) -> None:
         ...
