+++ stubs/micropython-v1_17-docstubs/select.pyi
@@ -1,12 +1,1 @@
-from typing import Any, Iterator, List, Optional, Tuple
-
-class poll:
-    def __init__(self) -> None: ...
-    def register(self, obj, eventmask: Optional[Any]) -> None: ...
-    def unregister(self, obj) -> Any: ...
-    def modify(self, obj, eventmask) -> None: ...
-    def poll(self, timeout: int = ...) -> List: ...
-    def ipoll(self, timeout: int = ..., flags: int = ...) -> Iterator[Tuple]: ...
-
-class select:
-    def __init__(self, rlist, wlist, xlist, timeout: Optional[Any]) -> None: ...
+from uselect import *
