+++ stubs/micropython-v1_17-docstubs/ssl.pyi
@@ -1,8 +1,1 @@
-from typing import Any
-
-SSLError: Any
-CERT_NONE: Any
-CERT_OPTIONAL: Any
-CERT_REQUIRED: Any
-
-def wrap_socket(sock, server_side: bool = ..., keyfile: Any | None = ..., certfile: Any | None = ..., cert_reqs: Any | None = ..., ca_certs: Any | None = ..., do_handshake: bool = ...) -> Any: ...
+from ussl import *
