+++ stubs/micropython-v1_17-docstubs/struct.pyi
@@ -1,7 +1,1 @@
-from typing import Any, Tuple
-
-def calcsize(fmt) -> int: ...
-def pack(fmt, v1, v2, *args) -> bytes: ...
-def pack_into(fmt, buffer, offset, v1, v2, *args) -> Any: ...
-def unpack(fmt, data) -> Tuple: ...
-def unpack_from(fmt, data, offset: int = ...) -> Tuple: ...
+from ustruct import *
