+++ stubs/micropython-v1_17-docstubs/uarray.pyi
@@ -1,0 +1,6 @@
+from typing import Any, Optional
+
+class array:
+    def __init__(self, typecode, iterable: Optional[Any]) -> None: ...
+    def append(self, val) -> Any: ...
+    def extend(self, iterable) -> Any: ...
