+++ stubs/micropython-v1_17-docstubs/ubinascii.pyi
@@ -1,0 +1,6 @@
+from typing import Any, Optional
+
+def hexlify(data, sep: Optional[Any]) -> bytes: ...
+def unhexlify(data) -> bytes: ...
+def a2b_base64(data) -> bytes: ...
+def b2a_base64(data) -> bytes: ...
