+++ stubs/micropython-v1_17-docstubs/uio.pyi
@@ -1,0 +1,17 @@
+import abc
+from typing import Any, IO, Optional
+
+class FileIO(IO, metaclass=abc.ABCMeta):
+    def __init__(self, *args) -> None: ...
+
+class TextIOWrapper:
+    def __init__(self, *args) -> None: ...
+
+class StringIO(IO, metaclass=abc.ABCMeta):
+    def __init__(self, string: Optional[Any]) -> None: ...
+
+class BytesIO(IO, metaclass=abc.ABCMeta):
+    def __init__(self, string: Optional[Any]) -> None: ...
+    def getvalue(self) -> Any: ...
+
+def open(name, mode: str = ..., **kwargs) -> Any: ...
