+++ stubs/micropython-v1_17-docstubs/ujson.pyi
@@ -1,0 +1,6 @@
+from typing import Any
+
+def dump(obj, stream, separators: Any | None = ...) -> Any: ...
+def dumps(obj, separators: Any | None = ...) -> str: ...
+def load(stream) -> Any: ...
+def loads(str) -> Any: ...
