+++ stubs/micropython-v1_17-docstubs/ussl.pyi
@@ -1,0 +1,8 @@
+from typing import Any
+
+SSLError: Any
+CERT_NONE: Any
+CERT_OPTIONAL: Any
+CERT_REQUIRED: Any
+
+def wrap_socket(sock, server_side: bool = ..., keyfile: Any | None = ..., certfile: Any | None = ..., cert_reqs=..., ca_certs: Any | None = ..., do_handshake: bool = ...) -> Any: ...
